Major changes: Hardened compilation requirements, documentation moved to doxygen.
org-based documentation has been removed (as has its infrastructure).

Code has had necessary alterations to compile with hardening/stricter
compile time checks with both GCC and Clang (hence all the awkward
static_cast-ing being done). This will be cleaned up later-on (much glue
is required to work with FLTK which uses less strict requirements).

Files are being updated to provide complete Doxygen documentation
generation and todos are being built-up (with in-code Doxygen comments)
for steps to take in the near future.

Updated the documentation generation script to reflect these
changes (make_docs.sh).

Updated format_lint.sh to include additional linting (analogous to what
is being done with sac-format).

+++ include/PsSp/Managers/SheetManager.hpp
@@ -1,4 +1,4 @@
-// Copyright 2023 Alexander R. Blanchette
+// Copyright 2023-2024 Alexander R. Blanchette
 
 #ifndef PSSP_SHEETMANAGER_HPP_20231219_1042
 #define PSSP_SHEETMANAGER_HPP_20231219_1042
@@ -25,24 +25,24 @@ namespace pssp {
 class SheetManager {
 public:
   SheetManager();
-  void resize_data(int size);
+  void resize_data(size_t size);
   [[nodiscard]] int rows() const;
   [[nodiscard]] int cols() const;
   // Setters
-  void set(int row, const Field &field, const std::string &input);
-  void set(int row, const Field &field, int input);
-  void set(int row, const Field &field, float input);
-  void set(int row, const Field &field, double input);
-  void set(int row, const Field &field, bool input);
+  void set(size_t row, const Field &field, const std::string &input);
+  void set(size_t row, const Field &field, int input);
+  void set(size_t row, const Field &field, float input);
+  void set(size_t row, const Field &field, double input);
+  void set(size_t row, const Field &field, bool input);
   // Getters
   // This does do conversion if necessary
-  std::string get(int row, const Field &field);
+  std::string get(size_t row, const Field &field);
   // These don't do conversions
-  std::string get_string(int row, const Field &field);
-  int get_int(int row, const Field &field);
-  float get_float(int row, const Field &field);
-  double get_double(int row, const Field &field);
-  bool get_bool(int row, const Field &field);
+  std::string get_string(size_t row, const Field &field);
+  int get_int(size_t row, const Field &field);
+  float get_float(size_t row, const Field &field);
+  double get_double(size_t row, const Field &field);
+  bool get_bool(size_t row, const Field &field);
 
 private:
   // cppcheck-suppress unusedStructMember
